[Security Solution][Flyout] add Table tab to new document flyout_v2
