+++ x-pack/platform/plugins/private/translations/translations/de-DE.json
@@ -43846,11 +43846,6 @@
     "xpack.securitySolution.flyout.threatIntelligence.refresh": "Aktualisieren",
     "xpack.securitySolution.flyout.threatIntelligence.requestError": "Beim Anfordern von Threat-Intelligence ist ein Fehler aufgetreten",
     "xpack.securitySolution.flyout.threatIntelligence.title": "Threat-Intelligence",
-    "xpack.securitySolution.flyout.tour.tableTab.title": "Tab „Tabelle anpassen“",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.description": "Pinnen Sie Felder an, die Sie interessieren oder die Sie für später speichern möchten.",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.title": "Felder anheften",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.description": "Ändern Sie die Einstellungen des Tabs, um bestimmte Felder anzuzeigen oder auszublenden. Beispielsweise können Sie auswählen, dass nur die hervorgehobenen Felder angezeigt werden.",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.title": "Felder ein- und ausblenden",
     "xpack.securitySolution.flyout.user.preview.viewDetailsLabel": "Vollständige Benutzerdetails anzeigen",
     "xpack.securitySolution.flyout.visualizations.analyzerPreview.noDataDescription": "Sie können nur Ereignisse visualisieren, die von Hosts ausgelöst wurden, die mit der Elastic Defend-Integration konfiguriert sind, oder beliebige {sysmon} Daten von {winlogbeat}. Weitere Informationen finden Sie unter {link}.",
     "xpack.securitySolution.flyoutV2.document.visualizations.analyzerPreview.coldAndFrozenTiers.excludedLabel": "Kalte/Eingefrorene Tiers {state}",
