+++ x-pack/platform/plugins/private/translations/translations/fr-FR.json
@@ -43696,11 +43696,6 @@
     "xpack.securitySolution.flyout.threatIntelligence.refresh": "Actualiser",
     "xpack.securitySolution.flyout.threatIntelligence.requestError": "Une erreur est survenue lors de la demande de Threat Intelligence",
     "xpack.securitySolution.flyout.threatIntelligence.title": "Threat Intelligence",
-    "xpack.securitySolution.flyout.tour.tableTab.title": "Personnalisez l'onglet Tableau",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.description": "Épinglez les champs qui vous intéressent ou que vous souhaitez enregistrer pour plus tard.",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.title": "Épingler les champs",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.description": "Modifiez les paramètres de l'onglet pour afficher ou masquer certains champs. Par exemple, choisissez de n’afficher que les champs surlignés.",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.title": "Afficher et masquer les champs",
     "xpack.securitySolution.flyout.user.preview.viewDetailsLabel": "Afficher tous les détails de l'utilisateur",
     "xpack.securitySolution.flyout.visualizations.analyzerPreview.noDataDescription": "Vous pouvez uniquement visualiser les événements déclenchés par les hôtes configurés avec l'intégration Elastic Defend ou les données {sysmon} provenant de {winlogbeat}. Pour en savoir plus, consultez {link}.",
     "xpack.securitySolution.flyoutV2.document.visualizations.analyzerPreview.coldAndFrozenTiers.excludedLabel": "Niveau froid/gelé {state}",
