+++ x-pack/platform/plugins/private/translations/translations/ja-JP.json
@@ -44012,11 +44012,6 @@
     "xpack.securitySolution.flyout.threatIntelligence.refresh": "更新",
     "xpack.securitySolution.flyout.threatIntelligence.requestError": "脅威インテリジェンスの要求中にエラーが発生しました",
     "xpack.securitySolution.flyout.threatIntelligence.title": "脅威インテリジェンス",
-    "xpack.securitySolution.flyout.tour.tableTab.title": "テーブルタブをカスタマイズしてください",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.description": "興味のあるフィールドや後で保存したいフィールドをピン留めします。",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.title": "フィールドをピン留め",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.description": "タブの設定を変更して、特定のフィールドを表示または非表示にします。例えば、ハイライトされたフィールドのみを表示するように選択します。",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.title": "フィールドの表示と非表示",
     "xpack.securitySolution.flyout.user.preview.viewDetailsLabel": "すべてのユーザー詳細を表示",
     "xpack.securitySolution.flyout.visualizations.analyzerPreview.noDataDescription": "Elastic Defend統合で構成されたホストまたは{winlogbeat}の{sysmon}データによってトリガーされたイベントのみを可視化できます。詳細については、{link}を参照してください。",
     "xpack.securitySolution.flyoutV2.document.visualizations.analyzerPreview.coldAndFrozenTiers.excludedLabel": "コールド/フローズンティア {state}",
