+++ x-pack/platform/plugins/private/translations/translations/zh-CN.json
@@ -44005,11 +44005,6 @@
     "xpack.securitySolution.flyout.threatIntelligence.refresh": "刷新",
     "xpack.securitySolution.flyout.threatIntelligence.requestError": "请求威胁情报时发生错误",
     "xpack.securitySolution.flyout.threatIntelligence.title": "威胁情报",
-    "xpack.securitySolution.flyout.tour.tableTab.title": "自定义 “表格” 选项卡",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.description": "固定您感兴趣的字段或想要保存以备后用的字段。",
-    "xpack.securitySolution.flyout.tour.tableTabPinning.title": "固定字段",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.description": "修改选项卡的设置以显示或隐藏某些字段。例如，选择只显示突出显示的字段。",
-    "xpack.securitySolution.flyout.tour.tableTabSetting.title": "显示和隐藏字段",
     "xpack.securitySolution.flyout.user.preview.viewDetailsLabel": "显示全部用户详情",
     "xpack.securitySolution.flyout.visualizations.analyzerPreview.noDataDescription": "您只能可视化由使用 Elastic Defend 集成或来自 {winlogbeat} 的任何 {sysmon} 数据配置的主机触发的事件。请参阅 {link} 了解更多信息。",
     "xpack.securitySolution.flyoutV2.document.visualizations.analyzerPreview.coldAndFrozenTiers.excludedLabel": "冷/冻结层{state}",
