+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/test_ids.ts
@@ -27,7 +27,6 @@ export const TABLE_TAB_SETTING_HIDE_EMPTY_FIELDS_TEST_ID =
   `${FLYOUT_TABLE_TEST_ID}HideEmptyFields` as const;
 export const TABLE_TAB_SETTING_HIDE_ALERT_FIELDS_TEST_ID =
   `${FLYOUT_TABLE_TEST_ID}HideAlertFields` as const;
-export const TABLE_TAB_TOUR_TEST_ID = `${FLYOUT_TABLE_TEST_ID}Tour` as const;
 
 /* Header */
