+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/attack/main/components/table_field_value_cell.tsx
@@ -13,7 +13,7 @@ import type { FieldSpec } from '@kbn/data-plugin/common';
 import { css } from '@emotion/react';
 import type { EventFieldsData } from '../../../../common/components/event_details/types';
 import { FormattedFieldValue } from '../../../../timelines/components/timeline/body/renderers/formatted_field';
-import { getFieldFormat } from '../../../../flyout/document_details/right/utils/get_field_format';
+import { getFieldFormat } from '../../../shared/utils/get_field_format';
 import { useExpandableValues } from '../../../../flyout/document_details/shared/hooks/use_expandable_values';
 
 export interface FieldValueCellProps {
