+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/main/components/table_tab_setting_button.tsx
@@ -23,7 +23,7 @@ import {
   TABLE_TAB_SETTING_HIGHLIGHTED_FIELDS_ONLY_TEST_ID,
   TABLE_TAB_SETTING_HIDE_EMPTY_FIELDS_TEST_ID,
   TABLE_TAB_SETTING_HIDE_ALERT_FIELDS_TEST_ID,
-} from './test_ids';
+} from '../../../../flyout/document_details/right/components/test_ids';
 
 const TABLE_TAB_SETTING_BUTTON_LABEL = i18n.translate(
   'xpack.securitySolution.flyout.documentDetails.right.tableTabSettingButton.label',
