+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/main/constants/local_storage.ts
@@ -11,4 +11,5 @@ export const FLYOUT_STORAGE_KEYS = {
   THREAT_INTELLIGENCE_TIME_RANGE:
     'securitySolution.documentDetailsFlyout.threatIntelligenceTimeRange',
   PREVALENCE_TIME_RANGE: 'securitySolution.documentDetailsFlyout.prevalenceTimeRange',
+  TABLE_TAB_STATE: 'securitySolution.documentDetailsFlyout.tableTabState.v8.19',
 };
