+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/main/utils/table_tab_utils.tsx
