+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/shared/components/table_field_name_cell.test.tsx
@@ -12,7 +12,7 @@ import { TableFieldNameCell } from './table_field_name_cell';
 import {
   FLYOUT_TABLE_FIELD_NAME_CELL_ICON_TEST_ID,
   FLYOUT_TABLE_FIELD_NAME_CELL_TEXT_TEST_ID,
-} from './test_ids';
+} from '../../../flyout/document_details/right/components/test_ids';
 
 const mockDataType = 'date';
 const mockField = '@timestamp';
