+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/shared/utils/get_field_format.ts
